requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
requests>=2.14.2 # Apache-2.0
paho-mqtt>=1.1 # EPL-1.0
subunit2sql>=1.9.0 # Apache-2.0
SQLAlchemy>=1.0.10,!=1.1.5,!=1.1.6,!=1.1.7,!=1.1.8  # MIT
pandas>=0.20.1
tensorflow>=1.6.0 # Apache-2.0
PyMySQL>=0.8.0
